Initial platform layer for the cc2650 chip

* commit '9308a31ecec01c494fe08fcc5e9ef0e1746bf02f': (231 commits)
  fixed build error from using the VerifyOrExit macro
  moved to using 'VerifyOrExit' in cc2650 platform layer to match the rest of the codebase
  Enhance otPlatRandomGet API to provide random stream (#471)
  Implement Server Data Notify. (#457)
  Only send Child Update Request message when no more state changes are pending. (#476)
  Add check for CLI argument number (#465)
  Ncp: Update the RESET_hanlder() to ensure a single spinel reset message goes out (#456)
  NcpSpi: Ensure spi tx frame preparation and callbacks happen from the same tasklet. (#455)
  posix uart termios typo (#463)
  Fix bugs in retrieving child/router info using an RLOC16. (#475)
  Don't reset key sequence when setting the same master key (#468)
  fix AppendRoute() bug: LinkQualityOut is always 0 (#466)
  Set VERBOSE=1 to provide additional output on travis builds. (#461)
  Fix bugs when comparing partitions during parent selection. (#472)
  Implement a method for removing a neighbor. (#454)
  Fix default route issue (#451)
  Ncp: Post the tasklet `mUpdateChangedPropTask` from constructor for initial reset flag. (#453)
  Update ADDRESS_QUERY_MAX_RETRY_DELAY parameter value. (#450)
  All secured MLE messages must have Key ID Mode 2. (#449)
  Adding openthread API for Energy Scan operation (#448)
  ...

+++ src/core/common/logging.cpp
@@ -50,49 +50,49 @@ extern "C" {
  * @param[in]  aLength     Number of bytes in the buffer.
  *
  */
-static void DumpLine(otLogLevel aLogLevel, otLogRegion aLogRegion, const void *aBuf, const int aLength)
+static void DumpLine(otLogLevel aLogLevel, otLogRegion aLogRegion, const void *aBuf, const size_t aLength)
 {
     char buf[80];
     char *cur = buf;
 
-    snprintf(cur, sizeof(buf) - (cur - buf), "|");
+    snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), "|");
     cur += strlen(cur);
 
-    for (int i = 0; i < 16; i++)
+    for (size_t i = 0; i < 16; i++)
     {
         if (i < aLength)
         {
-            snprintf(cur, sizeof(buf) - (cur - buf), " %02X", ((uint8_t *)(aBuf))[i]);
+            snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), " %02X", ((uint8_t *)(aBuf))[i]);
             cur += strlen(cur);
         }
         else
         {
-            snprintf(cur, sizeof(buf) - (cur - buf), " ..");
+            snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), " ..");
             cur += strlen(cur);
         }
 
         if (!((i + 1) % 8))
         {
-            snprintf(cur, sizeof(buf) - (cur - buf), " |");
+            snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), " |");
             cur += strlen(cur);
         }
     }
 
-    snprintf(cur, sizeof(buf) - (cur - buf), " ");
+    snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), " ");
     cur += strlen(cur);
 
-    for (int i = 0; i < 16; i++)
+    for (size_t i = 0; i < 16; i++)
     {
         char c = 0x7f & ((char *)(aBuf))[i];
 
         if (i < aLength && isprint(c))
         {
-            snprintf(cur, sizeof(buf) - (cur - buf), "%c", c);
+            snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), "%c", c);
             cur += strlen(cur);
         }
         else
         {
-            snprintf(cur, sizeof(buf) - (cur - buf), ".");
+            snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), ".");
             cur += strlen(cur);
         }
     }
@@ -100,42 +100,42 @@ static void DumpLine(otLogLevel aLogLevel, otLogRegion aLogRegion, const void *a
     otPlatLog(aLogLevel, aLogRegion, "%s\n", buf);
 }
 
-void otDump(otLogLevel aLogLevel, otLogRegion aLogRegion, const char *aId, const void *aBuf, const int aLength)
+void otDump(otLogLevel aLogLevel, otLogRegion aLogRegion, const char *aId, const void *aBuf, const size_t aLength)
 {
-    int idlen = strlen(aId);
-    const int width = 72;
+    size_t idlen = strlen(aId);
+    const size_t width = 72;
     char buf[80];
     char *cur = buf;
 
     otPlatLog(aLogLevel, aLogRegion, "\n");
 
-    for (int i = 0; i < (width - idlen) / 2 - 5; i++)
+    for (size_t i = 0; i < (width - idlen) / 2 - 5; i++)
     {
-        snprintf(cur, sizeof(buf) - (cur - buf), "=");
+        snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), "=");
         cur += strlen(cur);
     }
 
-    snprintf(cur, sizeof(buf) - (cur - buf), "[%s len=%03d]", aId, aLength);
+    snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), "[%s len=%03zu]", aId, aLength);
     cur += strlen(cur);
 
-    for (int i = 0; i < (width - idlen) / 2 - 4; i++)
+    for (size_t i = 0; i < (width - idlen) / 2 - 4; i++)
     {
-        snprintf(cur, sizeof(buf) - (cur - buf), "=");
+        snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), "=");
         cur += strlen(cur);
     }
 
     otPlatLog(aLogLevel, aLogRegion, "%s\n", buf);
 
-    for (int i = 0; i < aLength; i += 16)
+    for (size_t i = 0; i < aLength; i += 16)
     {
         DumpLine(aLogLevel, aLogRegion, (uint8_t *)(aBuf) + i, (aLength - i) < 16 ? (aLength - i) : 16);
     }
 
     cur = buf;
 
-    for (int i = 0; i < width; i++)
+    for (size_t i = 0; i < width; i++)
     {
-        snprintf(cur, sizeof(buf) - (cur - buf), "-");
+        snprintf(cur, sizeof(buf) - static_cast<size_t>(cur - buf), "-");
         cur += strlen(cur);
     }
