+++ src/core/net/ip6_routes.cpp
@@ -80,11 +80,11 @@ ThreadError Routes::Remove(Route &aRoute)
     return kThreadError_None;
 }
 
-int Routes::Lookup(const Address &aSource, const Address &aDestination)
+int8_t Routes::Lookup(const Address &aSource, const Address &aDestination)
 {
-    int maxPrefixMatch = -1;
+    int8_t maxPrefixMatch = -1;
     uint8_t prefixMatch;
-    int rval = -1;
+    int8_t rval = -1;
 
     for (Route *cur = sRoutes; cur; cur = cur->mNext)
     {
@@ -100,21 +100,21 @@ int Routes::Lookup(const Address &aSource, const Address &aDestination)
             prefixMatch = cur->mPrefixLength;
         }
 
-        if (maxPrefixMatch > prefixMatch)
+        if (maxPrefixMatch > static_cast<int8_t>(prefixMatch))
         {
             continue;
         }
 
-        maxPrefixMatch = prefixMatch;
+        maxPrefixMatch = static_cast<int8_t>(prefixMatch);
         rval = cur->mInterfaceId;
     }
 
     for (Netif *netif = Netif::GetNetifList(); netif; netif = netif->GetNext())
     {
         if (netif->RouteLookup(aSource, aDestination, &prefixMatch) == kThreadError_None &&
-            prefixMatch > maxPrefixMatch)
+            static_cast<int8_t>(prefixMatch) > maxPrefixMatch)
         {
-            maxPrefixMatch = prefixMatch;
+            maxPrefixMatch = static_cast<int8_t>(prefixMatch);
             rval = netif->GetInterfaceId();
         }
     }
